Fixed the postgres dependencies on volumes
Because the `db` service was also dependent on local filesystem, the
files were not getting copied and the db container was not inserting
the rows into the database. Accordingly, this commit adds a new
Dockerfile.db file.

Additionally, to enable the `restart` configuration in the
docker-compose.yml file, I upgraded the file to Version 2.